% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_LER.R
\name{calculate_LER}
\alias{calculate_LER}
\title{Calculate leaf elongation rate}
\usage{
calculate_LER(
  leaf_length_data,
  n_LER_for_mean = 2,
  output = c("means", "tidy_LER", "wide_LER")
)
}
\arguments{
\item{leaf_length_data}{data.frame or tbl / tbl (tibble) containing the leaf length data.}

\item{n_LER_for_mean}{specifies the number of calculated LERs which must be used to calculate
the mean LER.}

\item{output}{specifies what the the funtion should return. By default, a tibble of the calculated
mean LERs is returned ("means"). When selecting "tidy_LER", a tidy tibble containing all calculated
LERs will be returned. "wide_LER" will return a wide, more human readable tibble of the calculated LERs.}
}
\value{
A tibble containing LERs. Depending on specified output parameter, LER output varies.
}
\description{
Calculates the leaf elongation rate (LER) for each plant in the provided dataset.
}
\details{
This function takes in leaf growth data, where the first column are
plant_id's and the following columns are leaf length measurements, where the
column headers are in the following date-time format: yyyy/mm/dd hh:mm or
yyyy/mm/dd hh:mm:ss
The provided leaf growth data must be in a data.frame or tibble. By default,
it uses the two first LERs are used for the calculation of the LER mean. This
means that at least three leaf length measurements are needed, allowing a
minimum of two growth intervals to be calculated.
}
\author{
Jonas Bertels
}
