% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_needed_by_kinematic_analysis.R
\name{cumm_cell_no}
\alias{cumm_cell_no}
\title{Functions, not needed by the user, but needed by the script}
\usage{
cumm_cell_no(size_vect, pos_vect)
}
\arguments{
\item{size_vect}{is a vector containing all the cell sizes.}

\item{pos_vect}{is a vector containing all the positions at which the cell sizes were taken.}

\item{cumm_cell_no}{is a vector containing all cummulative cell numbers.}
}
\description{
Contains functions which are used by the \link{kinematic_analysis} function.
}
\author{
Jonas Bertels
}
