% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_fitted_cell_lengths.R
\name{get_all_fitted_cell_lengths}
\alias{get_all_fitted_cell_lengths}
\title{Get all fitted cell lengths: Calculates for each plant all cell lengths}
\usage{
get_all_fitted_cell_lengths(
  cell_length_data,
  interval_in_cm = 0.1,
  bw_multiplier = 1,
  alternative_bw = 0.5,
  tidy_cell_lengths = TRUE
)
}
\arguments{
\item{cell_length_data}{data.frame or tbl / tbl_df (tibble) containing the cell length data.}

\item{interval_in_cm}{specifies the inteval at which cell lengths should be calculated.}

\item{bw_multiplier}{multiplies the calculated bandwidth with the provided number.}

\item{alternative_bw}{sets an alternative bandwidth if no band width could be calculated for the plant.}

\item{tidy_cell_lengths}{\code{TRUE} by default. When \code{FALSE}, the function will return cell lengths as a human readable table.}
}
\value{
A tidy tibble, containing all fitted cell lengths for each plant.
}
\description{
Produces, for each plant, all cell fitted lenghts over for a certain interval.
}
\details{
This function takes in cell length data. The provided cell length data must
be in a data.frame or tbl / tbl_df (tibble).
The first column are the plant IDs and are used in the LER analysis (header =
plant_id) .
The second column contains the positions of the cell length measurements in
the leaf's growth zone (in cm) (header = position).
The third column contains the cell length measurements (header = cell_length).

Using the interval_in_cm variable, you indicate how often (i.e. at which
interval) a cell length should be calculated over the entire growth zone. By
default, it is set to 0.1 cm.

The bw_multiplier can be used to multiply the bandwidth. Usually, the
calculated bandwidth is OK to use. Multiplying it by a number between 0 and 1
will follow the profile more strict. Multiplying the bandwidth with a number
larger than 1 will make the fit more loose. Making the multiplier to small
could result in an error.

An alternative bandwidth can be specified if, for some plants, no bandwidth
could be estimated. Our advise is to use the mean bandwidth of bandwidths
which were calculated. You can use the
\link{get_pdf_with_cell_length_fit_plots} function to extract all these
bandwidths and use them to calculate the mean bandwidth. When not specified,
the alternative bandwidth is set to 0.5 Normally, the cell lengths are
returned in a tidy format. When tidy_cell_lengths is set to \code{FALSE}, a
human readable wide table is returned.
}
\author{
Jonas Bertels
}
