% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdf_with_cell_length_fit_plots.R
\name{get_pdf_with_cell_length_fit_plots}
\alias{get_pdf_with_cell_length_fit_plots}
\title{Cell lengths: Make a pdf containing the cell length fits.}
\usage{
get_pdf_with_cell_length_fit_plots(
  cell_length_data,
  interval_in_cm = 0.1,
  bw_multiplier = 1,
  output_bw_tibble = FALSE
)
}
\arguments{
\item{cell_length_data}{data.frame or tbl / tbl_df (tibble) containing the cell length data.}

\item{interval_in_cm}{specifies the inteval at which cell lengths should be calculated.}

\item{bw_multiplier}{multiplies the calculated bandwidth with the provided number.}

\item{output_bw_tibble}{\code{FALSE} by default. When \code{TRUE}, the function will return all calculated bandwidths as a tibble.}
}
\value{
Normally, only a pdf is produced in the work directory. See additional output when using \code{output_bw_tibble}.
}
\description{
Produces a pdf file in the work directory, containing all the graphs of the
fitted cell lengths and their derivatives.
}
\details{
This function takes in cell length data. The provided cell length data must
be in a data.frame or tbl / tbl_df (tibble).
The first column are the plant IDs and are used in the LER analysis (header =
plant_id) .
The second column contains the positions of the cell length measurements in
the leaf's growth zone (in cm) (header = position).
The third column contains the cell length measurements (header = cell_length).

Important: All cell length measurements are thus entered directly underneath
eachother.

Concerning the variables:
Using the interval_in_cm variable, you indicate how often (i.e. at which
interval) a cell length should be calculated over the entire growth zone. By
default, it is set to 0.1 cm.

The bw_multiplier can be used to multiply the bandwidth. Usually, the
calculated bandwidth is OK to use. Multiplying it by a number between 0 and 1
will follow the profile more strict. Multiplying the bandwidth with a number
larger than 1 will make the fit more loose. Making the multiplier to small or
to big could result in an error.

The function will always produce a pdf file with plots in the working
directory, also indicating the multiplier in the file name. If you set
output_bw_tibble to \code{TRUE}, then the function will also return all
calculated bandwidths as a tibble.
}
\author{
Jonas Bertels
}
