% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematic_analysis.R
\name{kinematic_analysis}
\alias{kinematic_analysis}
\title{Perform the kinematic analysis.}
\usage{
kinematic_analysis(LER_means, tidy_cell_lengths, meristem_size_micrometre)
}
\arguments{
\item{LER_means}{contains a tidy tibble with the previoulsy calculated mean leaf elongation rates for each plant.}

\item{tidy_cell_lengths}{contains a tidy tibble with the previously fitted cell lengths for each plant.}

\item{meristem_size_micrometre}{contains a tidy tibble with the meristem sizes in micrometre for each plant.}
}
\value{
A tidy tibble, containing all calculated kinematic parameters.
}
\description{
Combines calculated leaf elongation rates, fitted cell sizes and meristem
sizes to perform a kinematic analysis, resulting in data about cell division
and cell elongation.
}
\details{
This function takes in the data generated by \link{calculate_LER},
\link{get_all_fitted_cell_lengths} and a data.frame or tibble containing the
meristem size of each plant involved in the analysis. The first column of the
meristem size file should contain plant_id's and should be named exactly
plant_id. The second column should contain meristem sizes in micrometre and
should be named exactly mer_length_um. It are these column names that the
function will look for and use to select the data.
}
\author{
Jonas Bertels
}
